+++ mod_pow.h
@@ -2,11 +2,11 @@
 #define MOD_POW
 
 //#define ull unsigned long long
-typedef unsigned long long ull;
+typedef long long int ull;
 
 int mod_pow_recursive(int x, long y, int m);
 
-int mod_pow(ull x, ull y, ull m);
+ull mod_pow(ull x, ull y, ull m);
 
 int mod_fibonachi(unsigned long n, int m);
